+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [#3030](https://github.com/influxdb/influxdb/pull/3030): Fix excessive logging of shard creation.
 - [#3038](https://github.com/influxdb/influxdb/pull/3038): Don't check deleted shards for precreation. Thanks @vladlopes.
 - [#3033](https://github.com/influxdb/influxdb/pull/3033): Add support for marshaling `uint64` in client.
+- [#3090](https://github.com/influxdb/influxdb/pull/3090): Remove database from TSDB index on DROP DATABASE.
 
 ## v0.9.0 [2015-06-11]